Move /extract command to base.py
